+++ testingdir/test
@@ -1,0 +1,6 @@
+# this is a markdown file.. this is why we do this..
+
+
+> should work
+
+CREATE | WRITE, None ---> update, newfile
