+++ testingdir/test.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+int main(void) {
+  return 0;
+}
+
